+++ index.js
@@ -8,6 +8,7 @@ module.exports = {
 		ENACT_PACK_NO_ANIMATION: true
 	},
 	env: {
+		worker: true,
 		es6: true, // Enables ES6 globals
 		'shared-node-browser': true, // restrict to common globals to preserve isomorphic support
 		commonjs: true
@@ -93,7 +94,10 @@ module.exports = {
 			builtinGlobals: true,
 			hoist: 'all',
 			allow: [
-				'context'
+				'context',
+				'location',
+				'name',
+				'Notification'
 			]
 		}],
 		'no-throw-literal': 'error',